Code review suggestions
