+++ wp-irving.php
@@ -21,9 +21,6 @@ define( 'WP_IRVING_VERSION', '1.0' );
 register_activation_hook( __FILE__, 'flush_rewrite_rules' );
 register_deactivation_hook( __FILE__, 'flush_rewrite_rules' );
 
-// Utilities.
-require_once WP_IRVING_PATH . '/inc/utils/html-markup.php';
-
 // Base classes.
 require_once WP_IRVING_PATH . '/inc/endpoints/class-endpoint.php';
 
@@ -63,7 +60,6 @@ require_once WP_IRVING_PATH . '/inc/rewrites.php';
 
 // Template loading.
 require_once WP_IRVING_PATH . '/inc/templates/admin-bar.php';
-require_once WP_IRVING_PATH . '/inc/templates/head-management.php';
 require_once WP_IRVING_PATH . '/inc/templates/namespace.php';
 require_once WP_IRVING_PATH . '/inc/templates/site-theme.php';
